@@ -17,35 +17,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "^7.3.9",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
+    "@angular/animations": "^8.0.3",
+    "@angular/cdk": "^8.0.1",
+    "@angular/cli": "^8.0.6",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "^8.0.3",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/forms": "^8.0.3",
     "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "^8.0.3",
+    "@angular/router": "^8.0.3",
     "@auth0/angular-jwt": "^2.1.0",
     "@ngrx/effects": "^7.4.0",
     "@ngrx/entity": "^7.4.0",
     "@ngrx/router-store": "^7.4.0",
     "@ngrx/store": "^7.4.0",
     "@ngrx/store-devtools": "^7.4.0",
-    "@swimlane/ngx-charts": "^11.1.0",
-    "@toverux/ngx-sweetalert2": "^4.0.0",
+    "@sweetalert2/ngx-sweetalert2": "^5.1.0",
+    "@swimlane/ngx-charts": "^11.2.0",
     "async": "^2.6.2",
-    "body-parser": "^1.18.3",
+    "body-parser": "^1.19.0",
     "bootstrap": "^4.3.1",
-    "core-js": "^2.6.4",
+    "core-js": "^2.6.9",
     "cors": "*",
-    "d3": "^5.9.2",
-    "email-templates": "^5.0.5",
-    "express": "^4.17.0",
+    "d3": "^5.9.7",
+    "email-templates": "^5.1.0",
+    "express": "^4.17.1",
     "hammerjs": "^2.0.8",
     "jsonwebtoken": "^8.5.1",
     "multer": "^1.4.1",
@@ -53,20 +53,20 @@
     "passport": "*",
     "passport-jwt": "^4.0.0",
     "pg": "^7.11.0",
-    "pg-hstore": "^2.3.2",
-    "rxjs": "~6.3.3",
+    "pg-hstore": "^2.3.3",
+    "rxjs": "~6.5.2",
     "sequelize": "^5.9.0",
-    "sequelize-cli": "^5.4.0",
+    "sequelize-cli": "^5.5.0",
     "sweetalert2": "^7.33.1",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/language-service": "~7.2.15",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular/language-service": "^8.0.3",
     "@ngrx/schematics": "^7.4.0",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@types/node": "^8.10.49",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -79,6 +79,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.5"
   }
 }
