@@ -7,8 +7,7 @@ import { DragDropModule } from '@angular/cdk/drag-drop';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 
-import { SweetAlert2Module } from '@toverux/ngx-sweetalert2';
-
+import { SweetAlert2Module } from '@sweetalert2/ngx-sweetalert2';
 
 import { PlannerRoutingModule } from './planner-routing.module';
 import { SharedModule } from '@/_shared/modules';
